// src/types/colors.interface.ts
export interface ILoggerColors {
    red: string
    green: string
    yellow: string
    blue: string
    magenta: string
    cyan: string
    white: string
    reset: string
    black: string
    lightgray: string
    default: string
    darkgray: string
    lightred: string
    lightgreen: string
    lightyellow: string
    lightblue: string
    lightmagenta: string
    lightcyan: string
}